harmonize logs when dropping connections for abuse or protocol violation
Those will be a warning level log in the net.failconn category, including
the offending IP, and the "dropping connection" string. This should make
it easier to use fail2ban with monero.

Note that connections dropped for more innocuous reasons (such as already
having the maximum number of incoming connections) do not generate such
a log line (they will typically include "dropping connections", but will
not be in the net.connfail category.